Update query API to remove use of long-lived threads.
The query API now is a one-off function, rather than something
intended to be run in a thread. The user can request a
range of timestamps to be searched and any results to be reported
immediately once a certain requested quota of bytes is exceeded.

The actual range of timestamps searched (if the quota limit is
reached) will be returned from the query function together with
the results. The client will need to use this information to
initiate a new search, if the returned timestamp range does not
included the desired results.

The stress tests have also been updated accordingly.

The Libnetdata web buffer BUFFER is now used to store the results.